+++ src/models/SolidContainerModel.ts
@@ -1,5 +1,4 @@
 import {
-    arrayUnique,
     stringToSlug,
     urlDirectoryName,
     urlParentDirectory,
@@ -7,37 +6,59 @@ import {
     uuid,
 } from '@noeldemartin/utils';
 import { FieldType } from 'soukai';
+import { findContainerRegistration } from '@noeldemartin/solid-utils';
 import type { IModel , MultiModelRelation } from 'soukai';
 
-import IRI from '@/solid/utils/IRI';
+import { SolidEngine } from '@/engines';
 
 import SolidContainerDocumentsRelation from './relations/SolidContainerDocumentsRelation';
 import SolidContainsRelation from './relations/SolidContainsRelation';
 
 import { SolidModel } from './SolidModel';
+import SolidTypeRegistration from './SolidTypeRegistration';
 import type SolidDocument from './SolidDocument';
+import type { SolidModelConstructor } from './inference';
 
 export default class SolidContainerModel extends SolidModel {
 
+    public static rdfsClasses = ['ldp:Container'];
+
     public static timestamps = false;
 
-    public static boot(name?: string): void {
-        const modelClass = this;
-
-        // Add container definitions.
-        modelClass.rdfsClasses = arrayUnique((modelClass.rdfsClasses ?? []).concat([IRI('ldp:Container')]));
-        modelClass.fields = modelClass.fields ?? {};
-        modelClass.fields['resourceUrls'] = {
+    public static fields = {
+        name: {
+            type: FieldType.String,
+            rdfProperty: 'rdfs:label',
+        },
+        resourceUrls: {
             type: FieldType.Array,
-            required: false,
-            rdfProperty: IRI('ldp:contains'),
-            items: {
-                type: FieldType.Key,
-            },
+            rdfProperty: 'ldp:contains',
+            items: FieldType.Key,
+        },
+    };
+
+    public static async fromTypeIndex<T extends SolidContainerModel>(
+        this: SolidModelConstructor<T>,
+        typeIndexUrl: string,
+        childrenModelClass: typeof SolidModel,
+    ): Promise<T | null> {
+        const engine = this.requireEngine();
+        const fetch = engine instanceof SolidEngine ? engine.getFetch() : undefined;
+        const containerRegistration = await findContainerRegistration(
+            typeIndexUrl,
+            childrenModelClass.rdfsClasses[0],
+            fetch,
+        );
+
+        if (!containerRegistration)
+            return null;
+
+        const attributes = {
+            url: containerRegistration.value('solid:instanceContainer'),
+            name: containerRegistration.value('rdfs:label'),
         };
 
-        // Boot model.
-        super.boot(name);
+        return this.newInstance(attributes, true);
     }
 
     public resourceUrls!: string[];
@@ -46,6 +67,20 @@ export default class SolidContainerModel extends SolidModel {
 
     public documentsRelationship(): MultiModelRelation {
         return new SolidContainerDocumentsRelation(this);
+    }
+
+    public async register(typeIndexUrl: string, childrenModelClass: typeof SolidModel): Promise<void> {
+        const typeRegistration = new SolidTypeRegistration({
+            forClass: childrenModelClass.rdfsClasses[0],
+            instanceContainer: this.url,
+        });
+
+        typeRegistration.mintUrl(typeIndexUrl, true, uuid());
+
+        await typeRegistration.withEngine(
+            this.requireEngine(),
+            () => typeRegistration.save(urlParentDirectory(typeIndexUrl)),
+        );
     }
 
     protected contains<T extends typeof SolidModel>(model: T): MultiModelRelation {